Rename basic.py to continuous.py
