+++ baybe/optimizers/continuous.py
